siban/requirements.txt
aliyun-python-sdk-core==2.13.19
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.10.1
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
crcmod==1.7
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
mysqlclient==1.4.6
oss2==2.12.1
pycryptodome==3.9.7
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
